+++ docs/examples/umdo/sampling/s_sobieski.py
@@ -71,7 +71,9 @@ scenario = UMDOScenario(
     uncertain_space,
     "Mean",
     formulation_name="MDF",
-    statistic_estimation_settings=Sampling_Settings(),
+    statistic_estimation_settings=Sampling_Settings(
+        estimate_statistics_iteratively=False, n_samples=20
+    ),
     maximize_objective=True,
     uncertain_design_variables={"x_2": "{}+u_x_2"},
 )
@@ -84,8 +86,8 @@ scenario.add_constraint("g_2", "Margin", factor=3.0)
 scenario.add_constraint("g_3", "Margin", factor=3.0)
 
 # %%
-# and execute it with a gradient-free optimizer:
-scenario.execute(algo_name="NLOPT_COBYLA", max_iter=100)
+# We execute this scenario using the gradient-based optimizer SLSQP:
+scenario.execute(algo_name="NLOPT_SLSQP", max_iter=100)
 
 # %%
 # Lastly,