+++ src/gemseo_umdo/formulations/_statistics/iterative_sampling/standard_deviation.py
@@ -19,7 +19,6 @@ from __future__ import annotations
 from typing import TYPE_CHECKING
 
 from numpy import array
-from numpy import nan_to_num
 from numpy import newaxis
 
 from gemseo_umdo.formulations._statistics.iterative_sampling.variance import Variance
@@ -31,24 +30,16 @@ if TYPE_CHECKING:
 class StandardDeviation(Variance):
     """Iterative estimator of the standard deviation."""
 
-    __var_jac: Variance
-    """The iterative estimator of the Jacobian of the variance."""
-
     def __init__(self) -> None:
         super().__init__()
-        self.__var_jac = Variance()
 
     def _get_estimation(self) -> RealArray:
         return array(self._estimator.getStandardDeviation())
 
     def compute_jacobian(self, value: RealArray, jac_value: RealArray) -> RealArray:
-        self.jac_shape = jac_value.shape
-        return nan_to_num(
-            self.__var_jac.compute_jacobian(value, jac_value).reshape(self.jac_shape)
+        self.estimate_statistic(value)
+        return (
+            super().compute_jacobian(value, jac_value)
             / self._get_estimation()[:, newaxis]
             / 2
-        ).ravel()
-
-    def reset(self, size: int) -> None:
-        super().reset(size)
-        self.__var_jac.reset(size)
+        )